test(invoices): split credit-note tests to a new file (file-length limit)
The new required-field coverage pushed test_attendee_invoice_service.py to
1010 lines, over the 1000-line cap. Move the self-contained
TestGenerateAttendeeCreditNote class to test_attendee_credit_note_service.py,
importing the shared invoice test helpers from the original module.
